+++ packages/compilers-types/.gitignore
@@ -1,0 +1,2 @@
+build/
+node_modules/
