+++ packages/lib-sourcify/test/Compilation/SolidityCompilation.spec.ts
@@ -4,11 +4,8 @@ import path from 'path';
 import fs from 'fs';
 import { SolidityCompilation } from '../../src/Compilation/SolidityCompilation';
 import { solc } from '../utils';
-import {
-  CompilationTarget,
-  Metadata,
-} from '../../src/Compilation/CompilationTypes';
-import { SolidityJsonInput } from '../../src/Compilation/SolidityTypes';
+import { CompilationTarget } from '../../src/Compilation/CompilationTypes';
+import { SolidityJsonInput, Metadata } from '@sourcify/compilers-types';
 import chaiAsPromised from 'chai-as-promised';
 
 use(chaiAsPromised);
